requirements.txt
Django>=5.0,<6.1
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
mysqlclient>=2.2,<2.3
django-cors-headers>=4.3,<4.4
django-filter>=24.2,<24.3
Pillow>=10.4,<11.0
qrcode>=7.4,<7.5
reportlab>=4.1,<4.2
openpyxl>=3.1,<3.2
python-dotenv>=1.0,<1.1
